pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btc_max_knowledge_agent"
version = "0.1.0"
description = "Bitcoin knowledge management and retrieval system"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "requests>=2.26.0",
    "python-dotenv>=0.19.0",
    "argon2-cffi>=25.0.0,<26.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "black>=21.0",
    "isort>=5.0.0",
    "mypy>=0.9.0",
    "pylint>=2.11.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["btc_max_knowledge_agent*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
